+++ public/index.html
@@ -18,7 +18,6 @@
         <div class="header-top-mobile">
           <div class="menu-toggle">
             <span></span>
-            <!-- Этот span будет анимироваться -->
             МЕНЮ
           </div>
